Finished Details Screen.

+++ lib/features/main_layout/domain/entities/movie_details_entity.dart
@@ -6,7 +6,7 @@ class MovieDetailsEntity {
   final String overview;
   final String posterPath;
   final String releaseDate;
-  final double popularity;
+  final int popularity;
   final double voteAverage;
   final List<Genre> genres;
